21/03/01 전성원3 서버와 연동

+++ app/src/main/java/com/guys_from_301/stock_game/data/GameSetDao.kt
@@ -20,10 +20,10 @@ interface GameSetDao {
     fun getSetWithId(id: Int): GameSet
 
     @Query("SELECT setmonthly FROM GameSet")
-    fun getSetMonthly(): Float
+    fun getSetMonthly(): Int
 
     @Query("SELECT setsalaryraise FROM GameSet")
-    fun getSetSalaryRaise(): Float
+    fun getSetSalaryRaise(): Int
 
     @Query("SELECT setgamelength FROM GameSet")
     fun getSetGameLength(): Int
